+++ src/test/java/model/game/ChessBoardTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
